+++ packages/pob-dependencies/lib/index.d.ts
@@ -1,38 +1,38 @@
 
 interface Dependencies {
-  '@babel/core': '7.4.0';
-  '@babel/preset-env': '7.4.2';
+  '@babel/core': '7.4.3';
+  '@babel/preset-env': '7.4.3';
   '@babel/preset-react': '7.0.0';
   '@commitlint/cli': '7.5.2';
   '@commitlint/config-conventional': '7.5.0';
-  '@pob/eslint-config': '29.0.1';
-  '@pob/eslint-config-babel': '29.0.1';
-  '@pob/eslint-config-babel-node': '29.0.1';
-  '@pob/eslint-config-node': '29.0.1';
-  '@pob/eslint-config-react': '29.0.1';
-  '@pob/eslint-config-typescript': '29.0.1';
-  '@pob/eslint-config-typescript-react': '29.0.1';
+  '@pob/eslint-config': '29.1.0';
+  '@pob/eslint-config-babel': '29.1.0';
+  '@pob/eslint-config-babel-node': '29.1.0';
+  '@pob/eslint-config-node': '29.1.0';
+  '@pob/eslint-config-react': '29.1.0';
+  '@pob/eslint-config-typescript': '29.1.0';
+  '@pob/eslint-config-typescript-react': '29.1.0';
   '@types/jest': '24.0.11';
-  '@typescript-eslint/eslint-plugin': '1.5.0';
-  '@typescript-eslint/parser': '1.5.0';
+  '@typescript-eslint/eslint-plugin': '1.6.0';
+  '@typescript-eslint/parser': '1.6.0';
   'babel-eslint': '10.0.1';
-  'babel-jest': '24.5.0';
+  'babel-jest': '24.7.1';
   'babel-preset-latest-node': '3.1.1';
-  'babel-preset-modern-browsers': '13.1.0';
-  'eslint': '5.15.3';
+  'babel-preset-modern-browsers': '14.0.0';
+  'eslint': '5.16.0';
   'husky': '1.3.1';
-  'jest': '24.5.0';
+  'jest': '24.7.1';
   'jest-junit-reporter': '1.1.0';
   'lerna': '3.13.1';
   'lint-staged': '8.1.5';
-  'pob-babel': '^23.7.1';
-  'pob-lcov-reporter': '^1.0.1';
-  'pob-release': '^4.9.0';
+  'pob-babel': '^23.8.0';
+  'pob-lcov-reporter': '^1.0.2';
+  'pob-release': '^4.9.1';
   'prettier': '1.16.4';
   'repository-check-dirty': '^1.0.2';
   'typedoc': '0.14.2';
-  'typescript': '3.3.4000';
-  'yarn-update-lock': '^0.3.3';
+  'typescript': '3.4.1';
+  'yarn-update-lock': '^0.3.4';
   'yarnhook': '0.4.0';
 }