+++ packages/babel-plugin-fix-class-properties-uninitialized/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.9](https://github.com/christophehurpeau/pob/compare/babel-plugin-fix-class-properties-uninitialized@1.1.8...babel-plugin-fix-class-properties-uninitialized@1.1.9) (2019-04-05)
+
+**Note:** Version bump only for package babel-plugin-fix-class-properties-uninitialized
+
+
+
+
+
 ## [1.1.8](https://github.com/christophehurpeau/pob/compare/babel-plugin-fix-class-properties-uninitialized@1.1.7...babel-plugin-fix-class-properties-uninitialized@1.1.8) (2019-03-21)
 
 **Note:** Version bump only for package babel-plugin-fix-class-properties-uninitialized