+++ packages/pob-babel/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [23.9.0](https://github.com/christophehurpeau/pob/compare/pob-babel@23.8.0...pob-babel@23.9.0) (2019-04-05)
+
+
+### Features
+
+* update deps ([aa5d8aa](https://github.com/christophehurpeau/pob/commit/aa5d8aa))
+
+
+
+
+
 # [23.8.0](https://github.com/christophehurpeau/pob/compare/pob-babel@23.7.1...pob-babel@23.8.0) (2019-03-21)