+++ packages/yarn-update-lock/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.5](https://github.com/christophehurpeau/pob/compare/yarn-update-lock@0.3.4...yarn-update-lock@0.3.5) (2019-04-05)
+
+**Note:** Version bump only for package yarn-update-lock
+
+
+
+
+
 ## [0.3.4](https://github.com/christophehurpeau/pob/compare/yarn-update-lock@0.3.3...yarn-update-lock@0.3.4) (2019-03-21)
 
 **Note:** Version bump only for package yarn-update-lock